elaborate a bit on docstrings
Trying to be careful not to add too much theoretical maintenance burden (this
three-day project is almost over, but I still want to leave it in a
theoretically continuable state).  I think what I've covered should be
reasonably easy to keep up-to-date with refactoring, and the code can speak for
itself regarding issues I didn't mention.